EOD_Data/pyproject.toml
[project]
name = "eod-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
    "ta-lib>=0.6.3",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "azure-identity>=1.20.0",
    "azure-storage-blob>=12.24.1",
    "sqlalchemy>=2.0.38",
    "sqlalchemy-utils>=0.41.2",
    "pyodbc>=5.2.0",
    "fastapi>=0.115.11",
    "uvicorn>=0.34.0",
]
